+++ tdrs-backend/tdpservice/search_indexes/parsers/tanf_parser.py
@@ -4,177 +4,14 @@ import logging
 from ..models import T1  # , T2, T3, T4, T5, T6, T7, ParserLog
 # from django.core.exceptions import ValidationError
 from .util import get_record_type
+from .schema_defs.tanf import t1_schema
 
 logger = logging.getLogger(__name__)
 logger.setLevel(logging.DEBUG)
 
-class Field:
-    """Provides a mapping between a field name and its position."""
-
-    def __init__(self, name, length, start, end, type):
-        self.name = name
-        self.length = length
-        self.start = start
-        self.end = end
-        self.type = type
-
-    def create(self, name, length, start, end, type):
-        """Create a new field."""
-        return Field(name, length, start, end, type)
-
-    def __repr__(self):
-        """Return a string representation of the field."""
-        return f"{self.name}({self.start}-{self.end})"
-
-class RowSchema:
-    """Maps the schema for data rows."""
-
-    def __init__(self):  # , section):
-        self.fields = []
-        # self.section = section # intended for future use with multiple section objects
-
-    def add_field(self, name, length, start, end, type):
-        """Add a field to the schema."""
-        self.fields.append(
-            Field(name, length, start, end, type)
-        )
-
-    def add_fields(self, fields: list):
-        """Add multiple fields to the schema."""
-        for field, length, start, end, type in fields:
-            self.add_field(field, length, start, end, type)
-
-    def get_field(self, name):
-        """Get a field from the schema."""
-        return self.fields[name]
-
-    def get_field_names(self):
-        """Get all field names from the schema."""
-        return self.fields.keys()
-
-    def get_all_fields(self):
-        """Get all fields from the schema."""
-        return self.fields
-
-
-def active_t1(line, line_number):
-    """Parse row as active case data, T1 only."""
-    '''
-                    FAMILY CASE CHARACTERISTIC DATA
-
-    DESCRIPTION         LENGTH  FROM    TO  COMMENTS
-    RECORD TYPE         2       1       2   "T1" - SECTION 1
-    REPORTING MONTH     6       3       8   Numeric
-    CASE NUMBER         11      9       19  Alphanumeric
-    COUNTY FIPS CODE    3       20      22  Numeric
-    STRATUM             2       23      24  Numeric
-    ZIP CODE            5       25      29  Alphanumeric
-    FUNDING STREAM      1       30      30  Numeric
-    DISPOSITION         1       31      31  Numeric
-    NEW APPLICANT       1       32      32  Numeric
-    FAMILY MEMBERS      2       33      34  Numeric
-    TYPE OF FAMILY      1       35      35  Numeric
-    SUBSIDIZED HOUSING  1       36      36  Numeric
-    MEDICAL ASSISTANCE  1       37      37  Numeric
-    FOOD STAMPS         1       38      38  Numeric
-    FOOD STAMP AMOUNT   4       39      42  Numeric
-    SUB CHILD CARE      1       43      43  Numeric
-    AMT CHILD CARE      4       44      47  Numeric
-    AMT CHIILD SUPPORT  4       48      51  Numeric
-    FAMILY'S CASH       4       52      55  Numeric
-    CASH
-    AMOUNT              4       56      59  Numeric
-    NBR_MONTH           3       60      62  Numeric
-    TANF CHILD CARE
-    AMOUNT              4       63      66  Numeric
-    CHILDREN_COVERED    2       67      68  Numeric
-    NBR_MONTHS          3       69      71  Numeric
-    TRANSPORTATION
-    AMOUNT              4       72      75  Numeric
-    NBR_MONTHS          3       76      78  Numeric
-    TRANSITIONAL SERVICES
-    AMOUNT              4       79      82  Numeric
-    NBR_MONTHS          3       83      85  Numeric
-    OTHER
-    AMOUNT              4       86      89  Numeric
-    NBR_MONTHS          3       90      92  Numeric
-    REASON FOR & AMOUNT OF ASSISTANCE
-    REDUCTION
-    SANCTIONS AMT       4       93      96  Numeric
-    WORK REQ            1       97      97  Alphanumeric
-    NO DIPLOMA          1       98      98  Alphanumeric
-    NOT IN SCHOOL       1       99      99  Alphanumeric
-    NOT CHILD SUPPORT   1       100     100 Alphanumeric
-    IRP FAILURE         1       101     101 Alphanumeric
-    OTHER SANCTION      1       102     102 Alphanumeric
-    PRIOR OVERPAYMENT   4       103     106 Alphanumeric
-    TOTAL REDUC AMOUNT  4       107     110 Alphanumeric
-    FAMILY CAP          1       111     111 Alphanumeric
-    LENGTH OF ASSIST    1       112     112 Alphanumeric
-    OTHER, NON-SANCTION 1       113     113 Alphanumeric
-    WAIVER_CONTROL_GRPS 1       114     114 Alphanumeric
-    TANF FAMILY
-    EXEMPT TIME_LIMITS  2       115     116 Numeric
-    CHILD ONLY FAMILY   1       117     117 Numeric
-    BLANK               39      118     156 Spaces
-    '''
-
-    # all of the below is assumed T1, we will eventually expand this for other sections
-    # via get_record_type(line)
-
-    # split line into fields
-    family_case_schema = RowSchema()
-    family_case_schema.add_fields(
-        [  # does it make sense to try to include regex (e.g., =r'^T1$')
-            ('record_type', 2, 1, 2, "Alphanumeric"),
-            ('reporting_month', 6, 3, 8, "Numeric"),
-            ('case_number', 11, 9, 19, "Alphanumeric"),
-            ('county_fips_code', 3, 20, 22, "Numeric"),
-            ('stratum', 2, 23, 24, "Numeric"),
-            ('zip_code', 5, 25, 29, "Alphanumeric"),
-            ('funding_stream', 1, 30, 30, "Numeric"),
-            ('disposition', 1, 31, 31, "Numeric"),
-            ('new_applicant', 1, 32, 32, "Numeric"),
-            ('family_size', 2, 33, 34, "Numeric"),
-            ('family_type', 1, 35, 35, "Numeric"),
-            ('receives_sub_housing', 1, 36, 36, "Numeric"),
-            ('receives_medical_assistance', 1, 37, 37, "Numeric"),
-            ('receives_food_stamps', 1, 38, 38, "Numeric"),
-            ('food_stamp_amount', 4, 39, 42, "Numeric"),
-            ('receives_sub_child_care', 1, 43, 43, "Numeric"),
-            ('child_care_amount', 4, 44, 47, "Numeric"),
-            ('child_support_amount', 4, 48, 51, "Numeric"),
-            ('family_cash_recources', 4, 52, 55, "Numeric"),
-            ('family_cash_amount', 4, 56, 59, "Numeric"),
-            ('family_cash_nbr_month', 3, 60, 62, "Numeric"),
-            ('tanf_child_care_amount', 4, 63, 66, "Numeric"),
-            ('children_covered', 2, 67, 68, "Numeric"),
-            ('child_care_nbr_months', 3, 69, 71, "Numeric"),
-            ('transportation_amount', 4, 72, 75, "Numeric"),
-            ('transport_nbr_months', 3, 76, 78, "Numeric"),
-            ('transition_services_amount', 4, 79, 82, "Numeric"),
-            ('transition_nbr_months', 3, 83, 85, "Numeric"),
-            ('other_amount', 4, 86, 89, "Numeric"),
-            ('other_nbr_months', 3, 90, 92, "Numeric"),
-            ('reduction_amount', 4, 93, 96, "Numeric"),
-            ('reduc_work_requirements', 1, 97, 97, "Numeric"),
-            ('reduc_adult_no_hs_diploma', 1, 98, 98, "Numeric"),
-            ('reduc_teen_not_in_school', 1, 99, 99, "Numeric"),
-            ('reduc_noncooperation_child_support', 1, 100, 100, "Numeric"),
-            ('reduc_irp_failure', 1, 101, 101, "Numeric"),
-            ('reduc_other_sanction', 1, 102, 102, "Numeric"),
-            ('reduc_prior_overpayment', 4, 103, 106, "Numeric"),
-            ('total_reduc_amount', 4, 107, 110, "Numeric"),
-            ('reduc_family_cap', 1, 111, 111, "Numeric"),
-            ('reduc_length_of_assist', 1, 112, 112, "Numeric"),
-            ('other_non_sanction', 1, 113, 113, "Numeric"),
-            ('waiver_control_grps', 1, 114, 114, "Numeric"),
-            ('tanf_family_exempt_time_limits', 2, 115, 116, "Numeric"),
-            ('tanf_new_child_only_family', 1, 117, 117, "Numeric"),
-            ('blank', 39, 118, 156, "Spaces"),
-        ]
-    )
-
+def active_t1_parser(line, line_number):
+    """Parse line in datafile as active case data, T1 only."""
+    family_case_schema = t1_schema()
     # create search_index model
     t1 = T1()
     content_is_valid = True
@@ -208,9 +45,6 @@ def active_t1(line, line_number):
         logger.warn('Content is not valid, skipping model creation.')
         return
 
-        '''
-        '''
-
     # try:
     # t1.full_clean()
     t1.save()
@@ -243,14 +77,11 @@ def parse(datafile):
     line_number = 0
     for raw_line in datafile:
         line_number += 1
-        # logger.debug('Parsing this line: "%s"', raw_line)
         if isinstance(raw_line, bytes):
-            # logger.info("Line is bytes, decoding to string...")
             raw_line = raw_line.decode()
         line = raw_line.strip('\r\n')
 
         record_type = get_record_type(line)
-        # logger.debug('Parsing as type %s this line: "%s"', record_type, line)
 
         if record_type == 'HE' or record_type == 'TR':
             # Header/trailers do not differ between types, this is part of preparsing.
@@ -263,7 +94,7 @@ def parse(datafile):
                 # should be added to parser log in #1354
                 continue
             else:
-                active_t1(line, line_number)
+                active_t1_parser(line, line_number)
         else:
             logger.warn("Parsing for type %s not yet implemented", record_type)
             continue
