+++ tdrs-backend/tdpservice/search_indexes/parsers/util.py
@@ -38,3 +38,51 @@ def get_record_type(row):
     else:
         logger.debug('No match for row: %s' % row)
         return None
+
+class Field:
+    """Provides a mapping between a field name and its position."""
+
+    def __init__(self, name, length, start, end, type):
+        self.name = name
+        self.length = length
+        self.start = start
+        self.end = end
+        self.type = type
+
+    def create(self, name, length, start, end, type):
+        """Create a new field."""
+        return Field(name, length, start, end, type)
+
+    def __repr__(self):
+        """Return a string representation of the field."""
+        return f"{self.name}({self.start}-{self.end})"
+
+class RowSchema:
+    """Maps the schema for data rows."""
+
+    def __init__(self):  # , section):
+        self.fields = []
+        # self.section = section # intended for future use with multiple section objects
+
+    def add_field(self, name, length, start, end, type):
+        """Add a field to the schema."""
+        self.fields.append(
+            Field(name, length, start, end, type)
+        )
+
+    def add_fields(self, fields: list):
+        """Add multiple fields to the schema."""
+        for field, length, start, end, type in fields:
+            self.add_field(field, length, start, end, type)
+
+    def get_field(self, name):
+        """Get a field from the schema."""
+        return self.fields[name]
+
+    def get_field_names(self):
+        """Get all field names from the schema."""
+        return self.fields.keys()
+
+    def get_all_fields(self):
+        """Get all fields from the schema."""
+        return self.fields
